crud member & create transaction

+++ CashierFormApp/Views/Components/ProductHandler.cs
@@ -58,7 +58,6 @@ namespace CashierFormApp.Views.Components
 
         private void ProductHandler_Load(object sender, EventArgs e)
         {
-            // Set txtCode to read-only if in Edit mode
             txtCode.ReadOnly = IsEditMode;
         }
 
@@ -82,7 +81,6 @@ namespace CashierFormApp.Views.Components
                 e.Handled = true;
             }
         }
-
         private void btnSave_Click(object sender, EventArgs e)
         {
             if (string.IsNullOrWhiteSpace(ProductName))
@@ -97,7 +95,7 @@ namespace CashierFormApp.Views.Components
             product.Code = txtCode.Text;
             product.Name = txtProduct.Text;
             product.Stock = Convert.ToInt32(txtStock.Text);
-            product.Price = Convert.ToDouble(txtPrice.Text);
+            product.Price = Convert.ToSingle(txtPrice.Text);
 
             int result = 0;
 
@@ -121,9 +119,6 @@ namespace CashierFormApp.Views.Components
                     this.Close();
                 }
             }
-
-
-
         }
 
     }
